tests/wpt/tests/html-aam/roles-contextual.tentative.html

<!doctype html>
<html>
<head>
  <title>Tentative: HTML-AAM Contextual-Specific Role Verification Tests</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/wai-aria/scripts/aria-utils.js"></script>
</head>
<body>

<!--
  New sectionheader and sectionfooter roles.
  See https://github.com/w3c/aria/pull/1931
-->
<nav>
  <footer data-testname="el-footer" aria-label="x" data-expectedrole="sectionfooter" class="ex">x</aside>
</nav>
<main>
  <footer data-testname="el-footer-ancestormain" data-expectedrole="sectionfooter" class="ex">x</footer>
</main>
<nav>
  <header data-testname="el-header" aria-label="x" data-expectedrole="sectionheader" class="ex">x</header>
</nav>
<main>
  <header data-testname="el-header-ancestormain" data-expectedrole="sectionheader" class="ex">x</header>
</main>

<script>
AriaUtils.verifyRolesBySelector(".ex");
</script>

</body>
</html>
