third_party/blink/web_tests/external/wpt/html-aam/roles-contextual.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] el-footer
  assert_equals: <footer data-testname="el-footer" aria-label="x" data-expectedrole="sectionfooter" class="ex">x\n</footer> expected "sectionfooter" but got "generic"
[FAIL] el-footer-ancestormain
  assert_equals: <footer data-testname="el-footer-ancestormain" data-expectedrole="sectionfooter" class="ex">x</footer> expected "sectionfooter" but got "generic"
[FAIL] el-header
  assert_equals: <header data-testname="el-header" aria-label="x" data-expectedrole="sectionheader" class="ex">x</header> expected "sectionheader" but got "generic"
[FAIL] el-header-ancestormain
  assert_equals: <header data-testname="el-header-ancestormain" data-expectedrole="sectionheader" class="ex">x</header> expected "sectionheader" but got "generic"
Harness: the test ran to completion.
